Added grid charge padding to initial allocation
In the grid2grid computation, the grid charges are
padded in all three dimensions to speed up convolution.
However, the padding was being done by allocating
a temporary array.

With this commit, the paddings are added during
initial allocation of grid charges. So there is
no temporary array in grid2grid function anymore.
Additionally, wrapping is also removed in restriction
step.